requirements.txt
pandas
paramiko
py7zr
fpdf2
openpyxl
questionary
tabulate
pexpect; sys_platform != 'win32'